Add Qt3D Studio Scene Explorer tool
This tool is not intended as a viewer, but rather as a testbed for the
features and APIs of Runtime 2.0. So far there is support for viewing a
scene, and a view of the Slides available for the presentation.  It's
possible to control which slide is the active slide, and change between
slides.  Of course there is still work to be done to make this
functional.

There is also a basic scene and property viewer.  These views however
are not yet live views.

Change-Id: I43fb9d889eb0f9e1c770222073809bd837b510cd
Reviewed-by: Christian Stromme <christian.stromme@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>