Add support for rebuilding materials on the fly
While uncommon and infrequent, sometimes materials have to be regenerated
due to property changes that lead to changing the list of active shadow
casters, SSAO status, etc.

This was not possible to implement before due to not having support for
regenerating the material and the underlying shaders. Support is now added
via rebuildModelMaterial().

This also highlights the need for more careful management of Qt3D node
parenting. Relying on the auto-parent-on-set-or-add behavior is convenient,
but care must be taken for objects that are reused: no cached QShaderProgram
or QParameter instance's lifetime must be tied to the QMaterial since the
QMaterial may now be destroyed at arbitrary times.

Task-number: QT3DS-301
Change-Id: Ie40b7d20f6658243f549d8ca0927b7ce1ae1b87e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>