Test-drive subpresentations via Q3DStudioWindow
There is no .uia parsing and related concepts in place yet. The window
will however from now on allow passing in additional .uip files.
These are treated as subpresentations and the Qt 3D scene building happens
accordingly.

This allows verifying that the framegraph is built and the actual content
is rendered as expected. (choose File->Add subpresentation in q3dsviewer,
behind the scenes this will start outputting that scene into a texture that
is not used for anything yet but is visibile in e.g. RenderDoc)

Task-number: QT3DS-368
Change-Id: Idacc6bba59c2dd757b44b6617dff31ccdaa91899
Reviewed-by: Andy Nichols <andy.nichols@qt.io>