Follow ShadowMapResolution changes
Add the missing default to the metadata xml as well. Likely an oversight
on the application side. While we are at it, add default values in C++
too for lights, like we do for other classes. This is not really necessary
for fields that have a default value in the metadata, but as we see the
defaults can come and go, potentially leading to uninitialized variables.
Better avoid this.

Task-number: QT3DS-299
Change-Id: I4253a5d0bd2d6120fd7c627d072459cc8c0dec28
Reviewed-by: Andy Nichols <andy.nichols@qt.io>