Add an in-scene profiling view
Do two consecutive double-clicks, taps, or press F12 (once the 3DS window has
focus(?)) to unleash the awesomeness.

Uses Dear Imgui. The important point here is being in-scene and lightweight,
meaning this will function as-is on mobile and embedded. None of the
alternatives (widgets, embed Qt Quick, etc.) can satisfy these criterias.

CPU and memory stats are only available on Windows at the moment.

There are several further ideas in the pipeline (framegraph statistics in
particular). Also, the Qt 3D profiling could be hooked up in some form eventually.

Note: This is not a replacement for qt3dsexplorer and similar. The goal is to
keep this focused on graphics. Some of the debug features of q3dsviewer
will get migrated though (like it is already done with the shadow toggle).

Task-number: QT3DS-324
Task-number: QT3DS-325
Change-Id: Iac85867c7e3f45aff51af445d028f15fa044a88d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>