Support recreating the shadow framegraph subtree on the fly
Shadow map resolution is the only property that can cause this in practice
and that value unlikely to change (esp. since now we mark it as
non-animatable to follow 3DS1)

Nonetheless the logic is more straightforward this way.

Fix also the depth-stencil buffer logic. Lights can reuse the same buffer
only when the desired size is the same. They need their own otherwise.

There are also some important lessons & patterns in here when it comes
to parenting QNodes.

Task-number: QT3DS-428
Change-Id: I7cfec5a8ee6c065b10b2dc6b37b89afa46ace020
Reviewed-by: Andy Nichols <andy.nichols@qt.io>