Add most of the progressive AA supports bits
The matrix jiggling is broken since working with the plain projection
matrix is wrong. Will need to find a way to do this differently.

In practice this means there will not be a visible AA effect on screen although
the progressive passes should run as expected.

No support for combining with MSAA yet.

Scenes with shadow map, depth texture, etc. passes may be problematic for
now. There are also some potential inefficiencies. And a potential crash on
resize (reason unknown). These will have to be addressed later.

Task-number: QT3DS-330
Change-Id: Iaf8b51066ff6284bad9d4a9739e8c8f0ab52ca14
Reviewed-by: Andy Nichols <andy.nichols@qt.io>