Strip copyright headers when processing shader includes
The copyright headers that are present in the effectlib shader code make
the generated shader code excessively long and makes the code hard to
read. So we strip off these headers.

Also to improve readability of generated code there are now markers at
the start and end of includes that note which file the block is from.

Change-Id: I8f08d3365717e1cbd7fe6b45f050270f60dcae90
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>