Add support for adv. layer blend modes via custom shaders
Plus the test scenes. Including non-fullscreen cases which 3DS1 cannot
handle (with its fallbacks). Such scenes are fully supported by this
implementation.

There is one problem: neither color burn nor color dodge looks correct.
However, that is also the case with 3DS1 (and we are using the same
shaders) That has to be fixed in a follow up later on.

Handling MSAA layers is left as a future exercise.

Relies on BlitFramebuffer fixes in qt3d/5.10 (and hopefully soon dev).

Also revamps and simplifies the layer clear color handling - the
original was unnecessarily convoluted.

Task-number: QT3DS-127
Change-Id: I3e6e2fdac9663add47c1a3115c1a4ee88366e70d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>