Fix directional lights with custom materials
The lighting logic for directional lighting is weird, so I made sure we
have the expected values in the lighting parameters.  This change makes
sure that the position property of the light has the directional data in
the format that the custom material shader expects.  The direction
property is not used by directional lights in the custom material
shader.

Change-Id: I6ad2bbe3c6aa85579fbac9e48ed843dd398f87c3
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>