Improve handling of Cameras
Previously we only created one layer camera at a time, and inserted it
into the correct part of the scene later. This however is problematic
when you have entities who are children of the camara.

Cameras are now created when we create the entity tree for a layer.
There is still only one active camera per layer at at time, but now each
QCamera is available and part of the scene.

Change-Id: I082853e0e9effddda082ce37048df1c4b52cf77f
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>