Q3DS Lancelot Graphics Regression test framework
Based off of the approach of the Qt Quick lancelot tests.
q3dsscenegrabber is responsible for running the tests and generating an
image. tst_q3ds is the lancelot test that submits images to the baseline
server.

The baseline code is copied from qtbase since we can't assume that we
are built under the qt5 directory structure. At some point in the future
the baseline code will likely be moved into its own repository, but for
now we can just take a copy.

Change-Id: I5ff86fc43679accae47a96da0e5a6363ceda99e0
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>