Split Q3DSWindow and introduce Q3DSEngine
Avoid tight coupling with QWindow. Instead, have a separate engine object
that controls all the rendering, mostly hiding Qt 3D, without any window
dependencies. Events like resize or input are expected to be passed to the
engine via handleXxxxEvent(). The QWindow or QOffscreenSurface is
communicated via setSurface(), which must be done before calling
setSource(). Mirroring the previous Q3DSWindow behavior, parsing and scene
building is done in setSource() but the Qt 3D rendering only starts upon
calling start(). When used in combination with Q3DSWindow this is done
automatically upon expose (as is setSource, and resize+input are forwarded
as well).

This is necessary in order to enable the creation of the Studio3D Quick
item, and make other future use cases (like rendering completely
offscreen) possible.

Q3DSWindow continues using the default behavior of taking the size from
the presentation upon first show, while sticking to the normal QWindow
size afterwards (e.g. when resizing interactively). When this is not
desired, and the presentation is to follow the window size right from the
start, call forceResize().

Task-number: QT3DS-836
Change-Id: I97c2f078291e482570c3bc82b276eb5751fbd778
Reviewed-by: Andy Nichols <andy.nichols@qt.io>