profile ui: Show something about shader programs
Including the error log (or at least the beinning of it) for shaders that
failed to compile. This can become handy with 3rd party custom materials
and effects.

Clean up the unused shaderPrefix concept a bit in the default and custom
material generators. Do not have unused strings but let the provided string
or the custom shader name reach ShaderProgramGenerator.

Change-Id: I6153b6ec3d77ef4923ee491f034742d0905e7637
Reviewed-by: Andy Nichols <andy.nichols@qt.io>