Studio3D: Stabilize
On Linux at least it is now able to start up and even change the window
both with the basic and threaded render loops. There are some issues
left in the threaded case, the window change can deadlock sometimes.
Also, in the threaded case the engine is not actually destructed when
exiting the application (since the deleteLater() request has no chance
to run). We can probably live with the latter for now. The former needs
more testing.

Change-Id: I7dac8a8b603973ce56cd8afee7075177aef4deda
Reviewed-by: Andy Nichols <andy.nichols@qt.io>