Add an event log window
Say hello to the log window, invokable either from the Scene info pane, or
by pressing ` (the latter should be an easy and familiar shortcut to
toggle; the window/item needs to have focus though).

The creation of the first Q3DSEngine will kick off capturing the Qt
debug/warning messages. The Qt message handler is still called. The
captured content is always pushed to the current main presentation's profiler
and is thus available in the log window (until cleared).

All this enables introducing performance hints, markers for interesting
events (e.g. material rebuilds or other unideal events) in the future
since the messages are available in one central place regardless of the
platform and the methods of launching the app.

Change-Id: I6b75e149482b66689dd3d7ba2ff8c2039866c739
Reviewed-by: Andy Nichols <andy.nichols@qt.io>