Add Q3DSSurfaceViewer API
Same disclaimer as for widgets, rendering is not perfect yet.

Offscreen readback is broken for reasons not known. (QT3DS-1042)

When rendering offscreen (and then reading back to generate images for
example), the global animation time cannot be controlled yet.
(to be handled in QT3DS-1041)

As a bonus, Qt 3D apparently does not like when the event loop is never
pupmped (because the application returns right away from main()) and
tends to deadlock upon cleanup. Worked around with a processEvents in
the example, needs a better solution later.

The API is revised somewhat when compared to 3DS1. Fixed frameUpdate
naming, now uses normal Qt patterns for create-destroy, etc.

Change-Id: Iee0c39483831b68fbc1889922d54b3cc5e8726e2
Reviewed-by: Andy Nichols <andy.nichols@qt.io>