Make the two uses of Q3DSPropertyChange distinct in the API
A Q3DSPropertyChangeList can be used in two ways:

1. passed both to applyPropertyChanges and notifyPropertyChanges
In this case both the property name and value must be set, obviously.

2. when passed only to notifyPropertyChanges
Here there is no value since the new values are expected to be already set
via some other mechanism (i.e. the static setters in case of animations).

As a first step towards cleaning up how values are get/set for graph
object members, make this distinction obvious in Q3dSPropertyChange.

Change-Id: I96afe73932ad272426b9cbc511b6f596a0e8ed87
Reviewed-by: Andy Nichols <andy.nichols@qt.io>