Revise how properties can be set on graph objects
1. Introduce proper setters

The pattern would be

Q3DSPropertyChange setSomething(TYPE newSomethingValue);

which would then allow writing the following, using starttime and
endtime as an example

obj->notifyPropertyChanges({ obj->setStartTime(5000), obj->setEndTime(6000) });

to both set and notify about the change. Pretty convenient.

2. Move the anim setters out of the way

The static, variant-based setters and getters used by the animation
system are moved to their own classes. (Q3DSnnnn -> Q3DSnnnnAnimator)
This makes the Q3DSnnnn classes a lot cleaner and the static stuff can also
be hidden later on, if needed (since they do not need to be public as long
as the friend declaration is there).

3. Custom property getters/setters for CustomMaterialInstance and EffectInstance

This should be self-explanatory:
    QVariantMap customProperties() const
    QVariant customProperty(const QString &name) const
    Q3DSPropertyChange setCustomProperty(const QString &name, const QVariant &value)
Additionally they both get a ctor that takes a Q3DSEffect or Q3DSCustomMaterial.

4. Make a bunch of functions on Q3DSUipPresentation and Q3DSSlide
public. This will not suffice for creating a full scene programatically
with animations and everything, but is a good start, as shown in the new
autotest.

Task-number: QT3DS-1123
Change-Id: I3a96fc8a38e7d483ab802bddf94999c68525b33f
Reviewed-by: Andy Nichols <andy.nichols@qt.io>