Fix handling of StopAtEnd in the slide player
The player made a distinction between "stopped" and "paused", assuming
that stop meant that the animations would be reset to the starting
point. This differs from how runtime 1.0 handles stop and StopAtEnd.
In addition to stop having different meaning, we would not properly
reset the animatables values to their initial values, as the values
from seeking and the rollback values could be update in the wrong
order, so we now do a forced reset when stop is called, to avoid getting
stale values from Qt3d.

Task-number: QT3DS-1145
Change-Id: I3c1cca6648faf3ede4bbed46189fe6bade4dff65
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>