Add utility function for getting the start and end time for a slide
Get the start and end time from the slide by looking at the parents
layer. This assumes that the properties for the layer(s) has been
updates before being called. If the parent doesn't have a layer, e.g.,
because it's a component, then we get the time by going through the
the start and endtime of the slide's objects.

Change-Id: Id8d2d134de3efd880ee5379b1316d5f1d3502f5a
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>