Move name() to Q3DSGraphObject and introduce object lookup by name
While not very straightforward at first, the concept of id and name is
like this:

1. all objects must have a unique ID. In the editor this is not exposed
on the UI and is handled internally. With programmatic scenes the id is
specified via newObject/registerObject (uniqueness is enforced).

2. name is the name that is shown in the object tree in the editor and
is freely changeable. Uniqueness is not a requirement (though the editor
shows a warning when using an already existing name).

3. there is a per-type default name in the data model metadata ("Scene",
"Layer", "Image", etc.) which is to be used when the source is a .uip
and no name attribute is provided explicitly.

In the runtime the name is not used in practice (object linking in .uip
is by id of course). Nonetheless, introduce object lookup by name, so
one can write

auto model = pres.objectByName<Q3DSModelNode>(QLatin1String("fancy name
specified in the editor UI"))

since, when using the editor, name is visible to the designer/developer,
while id is not.

Change-Id: I0a61289dee240794554d3349e21e17bdee3a3cc7
Reviewed-by: Andy Nichols <andy.nichols@qt.io>