Update node visibility based on starttime/endtime
Honor all items starttimes on initializing a slide to be shown

Honor the starttime/endtime for components and refactor/consolidate code

Refactors SlidePlayer so that the setSlideTime method can replace the
updateSlideVisibility method. Consolidates the two methods for taking
care to mark a slide and its children as visible/hidden. In addition,
the patch honors the starttime/endtime properities for components.

In order to do this it was necessary to change the SliderPlayer time API
(seek, position, duration) to use milliseconds instead of a normalized
range from 0 - 1.

I've also fixed bugs with positionChanged signal so that it consistently
emits only after the animation engine has registered the change and only
when the position has actually changed. This required a small change to
SlideExplorerWidget as well.

Fixes slideplayer test to take into account changes to (seek, position,
duration) indicated above and fixes a bug in the slideChanged signal
that was not being captured before.

Finally, I've added tests for component starttime/endtime visibility.

Task-number: QT3DS-123
Change-Id: Iebc8c0a6365d6551aec70f76aacffd7c00cd8ac4
Reviewed-by: Christian Stromme <christian.stromme@qt.io>