Change setMesh signature
...to take the name only so the 'static' and 'dynamic' setter APIs are
in sync and do not interfere with each other. Test this in the autotest
as well. There is no way to manually create a mesh (MeshList) at the moment
anyways.

This involves changing the setter implementation to prevent returning
early before we get a chance to do some custom actions.

Change-Id: Iafc499c12033203dbd9a1a5f82f3ae2a61585b6c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>