Add QML engine for behavior snippets
Does not include the Qt3DStudio.Behavior API implementation but the
interface from 1.0/1.1 is there. We now keep all import versions (1.0, 1.1)
as well

The QQmlEngine is always a dedicated one (for all behavior snippets).
This may be revisited later on to see if it possible to use the main
QQmlEngine when the application is using QML anyways.

While all behavior instances get the QML component and object created
upon loading the presentation, the Q3DSEngine API is prepared for editor
integration in the future since it is also possible to dynamically
insert/remove a Q3DSBehaviorInstance in the scene graph and call
(un)loadBehavior afterwards to kick off the QML component/object
creation/destruction.

Task-number: QT3DS-918
Change-Id: I72459233e0386ef85c6b0cd9f168da77c9751d54
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>