Fix potential crash on application exit
A previous commit changed the object ownership of the AnimationManager
class. The change did not result in a crash on exit, but that was by
luck. The animation manager is a shared class between the slideplayer
and its children. If the parent slideplayer deletes the animation
manager and the children are then destructed it could cause a crash on
exit in the future. This patch fixes it by sharing ownership of the AM
by all slideplayers.

Change-Id: Id9ae6b0459d035efc83b66befde089d0728cf19b
Reviewed-by: Christian Stromme <christian.stromme@qt.io>