Fix material diffuse color usage
Unlike 3DS1 we do not multiply it in the light's diffuse color and is
done in the shader instead. However, placing it under the
!hasEmissiveMap condition seems wrong: when having an emissive map
(barrel example), the material's diffuse color has to be taken account
in the same way. It does not make a difference with the default (1 1 1)
color but matters once it's changed to something else.

Change-Id: I6ae2e256bc70139aea61f361d82a9622686a5201
Reviewed-by: Andy Nichols <andy.nichols@qt.io>