Drop the warning when no camera in raycast
This is now verified to be normal after all. When changing slides and
the camera of a layer is not on the current (or master slide), the
result is null and we can safely ignore the layer in question.

Still, this does not explain why the first press+release is ignored after
the slide change for a layer that has its camera on the new slide as well.

Change-Id: Icd0026955d81cda20c9e41bf8f0cfe72d9bd0403
Reviewed-by: Andy Nichols <andy.nichols@qt.io>