Fix a bug with flickering component animations
The way we currently handle timeline visibility is a dummy callback from
the animationmanager to the main slideplayer that then iterates through
its objects (which can be component objects with their own slides) and
recursively calls SlidePlayer::setSlideTime. This is a problem for
independently animating a component while its parent slide is stopped or
paused for instance. This patch fixes a bug that causes flickering in
one of our test cases and references a TODO for enabling independent
control of the component.

Change-Id: I26f888c139f441ff9206e5e33ab66c5cb02f47f1
Reviewed-by: Christian Stromme <christian.stromme@qt.io>