Support using Alias Nodes
Alias nodes are weird, and end just just being copies in practice.
Anytime there is a Alias node in the scene, a copy of the referenced
node's sub-tree is copied and instanced.  This is recreating the
behavior in runtime 1.x, but is also not so useful because changes to
the original reference at runtime are not reflected by these references.

There is also some unusual behavior when it comes to the states defined
by the slide graph, because it is hard to define which states should be
used by the new instances.  This commit tries to recreate the behavior
of runtime 1.x, however this is difficult because there is significant
differences between how things look in the editor vs the runtime for 1.x
already.

Change-Id: I37446da8d0710ac43e9db2516f9c540201a51c5f
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>