Add support for EmitSignal in actions
This is just a way to notify the user application code about something
happening. Hook it up to a signal in Q3DSEngine (for private API users)
and Q3DSPresentation (public API). The node path is omitted since the
concept of 3DS1 horrible absolute paths is not desirable in 2.0 and
there is nothing one can do with it in the public API anyway.

Demonstrate it in the simpleqml example and actionevent.uip.

Change-Id: I3096ffc20ab5a401195a4da4e4f37208e571ab6c
Reviewed-by: Christian Stromme <christian.stromme@qt.io>