Q3DSPresentation public API: fireEvent and a few enablers
Some of the to-be-added APIs involve the silly element path concept which
we cannot avoid now. So move the lookup function to Q3DSEngine and add the
opposite that constructs a Scene.Layer.BlahBlah type of path from an object.

The duplication of overrides for Q3DSPresentationController functions is
now avoided by implemented some of the handleXxxxx functions in the base
class (the ones that just forward to Q3DSEngine).

Change-Id: I67cca6c0f56e545c1ba4dbf040dff26273bf823f
Reviewed-by: Christian Stromme <christian.stromme@qt.io>