Store opacity properties in UIPPresentation from 0 - 100
Opacity is registered in the UIP document as an value between 0-100.
Previously we were converting these in our scene data as a normalized
float 0-1 because thats how we will consume it later.  To avoid
unnecessariy conversion and complication lets continue to store the
original value, and only normalize it when we use it in the
scenemanager.

Change-Id: Ic7a44e5be4cd99abdb09b43dbf45063bcc9c4ff5
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>