Fix light probe logic glitch in def.mat. generator
It was enabling the ibl probe path even when the scene had nothing to do
with probes.

While we are at it, fix the features list to not use the m_ prefixes as
is the custom in such structs.

This fixes the occasional flicker in some scenes (esp. onces with
shadows).

Task-number: QT3DS-1327
Change-Id: I3feeeee53add7231e6d12cad441cad0fd6148571
Reviewed-by: Christian Stromme <christian.stromme@qt.io>