Q3DSPresentation: add profilingEnabled property
Have a way in the public API to enable the profile UI and to load
presentations with QObject tracking enabled. The private API separates
these two but for simplicity the public API has a single flag for both.

Defaults to disabled which is what applications need in production.
During development one can just do profilingEnabled: true.

Change-Id: I228d7936897b6891bf4727ade126b4dbda72f73f
Reviewed-by: Andy Nichols <andy.nichols@qt.io>