Add API for toggling profile UI and disable the implicit toggles
In the viewer profiling is enabled by default and the built-in toggles
(F10, backtick, double double click) or the menu item (when using
widgets) can be used to bring up the UI.

For applications none of these is ideal. Instead, provide a
Q3DSPresentation property that can be used to toggle the UI.

Therefore, an application will typically do profilingEnabled: true on
its Presentation and then set profileUiVisible to true when it wants to.
This way we are not tied to the sometimes not ideal keyboard and mouse
shortcuts in real applications.

Rename Q3DSEngine::profileUiEnabled to autoToggleProfileUi because
that's what it really is. It merely enables or disables the built-in
keyboard and mouse shortcuts for toggling, nothing more.

Change-Id: If49ed7bc2575b7e81c2f6101dba307a639c3c011
Reviewed-by: Andy Nichols <andy.nichols@qt.io>