Support Scoped Lights
It is possible to set the scope of any light node. The default behavior
when no scope is set to a light is to set the scope to the layer itself.
So previously we would just assume every light has layer scope and did
all light gathering on the layer level.

To support scoped lights it is necessary to build up a custom lighting
buffer each branch of the layer subtree, so the lighting structures have
been moved to Q3DSNodeAttached. Most of the time this is just a null
reference, but if a node has scoped lighting logic it will have an
alternative lighting buffer for its subtree items.

Shadow casting does not work when a light has a scope, but this does not
work with the 1.x runtime either.

Change-Id: Ib329daa6f3130de8a130fb2fe584261f71c89b86
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>