Add support for eyeball in effects
Testable via the console too, e.g. load up corona.uip and toggle with
set(Corona, eyeball, 0) and set(Corona, eyeball, 1)

This is also handy for more dynamicity in the future, when effect
instances can appear and disappear on the fly. Some enablers for
activating and deactivating on the go are now in place.

With MSAA-enabled layers special considerations are needed since the
shader program used to compose may need switching between the MSAA (no
effect) and the non-MSAA (with effect) variants. This is now in place as
well.

Task-number: QT3DS-1401
Change-Id: Ic3f4c10981e37854f2d96c2c4b91c0253d252ba6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>