Create QLayer objects with an entity parent
Massive stability fix. Parenting to a QNode into the framegraph leads to
issues with reloading or opening a new presentation in the viewer since
the objects (like those QLayer instances) get destroyed - due to being
parented in the framegraph hierarchy which itself is a subtree under the
root entity - while they may still be referenced from other parts of the
tree.

The comment about not wanting to parent a QLayer component to an entity
is plain wrong: specifying an entity as the parent does not
automatically add the component to that entity.

Change-Id: I413d75ac847b0ae3c287b33ac369e7d8975cf60e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>