Revise Qt 3D scene init
More precisely, the point where we attach the root entity to the aspect
engine. The behavior does not change when opening the viewer, widget,
etc. since this is still tied to the explicit start() call. However,
upon reloading or opening a new presentation we now defer setting the
new root until all (sub)presentations are created. This avoids subtle
rendering issues in scenes with subpresentations.

Change-Id: Ic2b9453363991008abc6e98ec81db290fa5a98b3
Reviewed-by: Andy Nichols <andy.nichols@qt.io>