Add Q3DSDataInput
Exposed as DataInput in the QML API.

Also fixes a parsing issue with controlledproperty attributes in the Logic
section. (namely that the attribute as not parsed if it was in an
Add/Set). This involves some reorganizing of how the data is collected
since we now need to accumulate both from the Scene and Logic sections.

Also fixes cloning (Alias node) of controlledproperty data.

Also adds a presentationLoaded() signal to Q3DSWidget and
Q3DSSurfaceViewer because applications may want to act whenever a new
scene is loaded (e.g. call setValue on some Q3DSDataInput object).
Studio3D handles this internally and DataInputs magically just work,
setting the values even when doing a reload or changing the source
filename.

Also adds a command 'datainput' to the console to allow quicker
investigation of data input connections in the scene.

Task-number: QT3DS-1404
Change-Id: Id2436490721e3386ba8604ca29ad4bda785521d2
Reviewed-by: Andy Nichols <andy.nichols@qt.io>