Force update on the animators when seeking
The tracking on the animators were removed in a previous change, so the
value the fronted node was holding was assumed unchanged, and thus not
causing the position to be updated.
Instead of enabling tracking again, which comes with an overhead, we
instead adjust the value slightly to trigger the update.

Task-number: QT3DS-1312
Change-Id: I629d73a5464ef89d9cee54bf3bb2c3994406cbbf
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>