Add a dummy ViewerSettings
Only showRenderStats is supported now. Other values are ignored.

This allows using QML code written for 1.x without having to remove the
ViewerSettings element. We may also support some of the other properties
later on, e.g. the scale mode (to be decided).

The design problems of the 1.x class are corrected, as usual: there is no
need for a private class here, uses Q_ENUM instead of Q_ENUMS, uses
argument-less  signals as is the custom with qml, etc.

Task-number: QT3DS-1521
Change-Id: I0a1ce763d263353876113f3675fc620767f8b966
Reviewed-by: Andy Nichols <andy.nichols@qt.io>