more API docs
Also pull in the property name table (the up-to-dateness of which is
questionable).

The version number in the generated docs should be 2.0, not the
Qt version. Will need to figure out later how to maintain this.

Not everything is documented yet. Less important functions will get
covered later.

Many bits here and there are revised compared to the 1.1 docs.
There are however chunks that have been taken as-is. These need to be
cleaned up later on.

Most importantly, an initial Getting Started page has been added. This,
like the mobile/embedded page needs to be extended signficiantly later on.

A set of unused constructors (the ones taking no presentation) are
removed from Q3DSElement/SceneElement/DataInput. This is because
the QML API does not use them while it is impossible to utilize these
in a meaningful way from C++ (where are require setting the presentation
up-front with no setter afterwards).

There were a few qdoc warnings from the internal implementations due to
useless /*! or backslashes that confused qdoc. These are removed.

Task-number: QT3DS-1606
Change-Id: I141bdc774e5be3e4fed5c1d729cac9b68a02be3b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>