Make calling Q3DS::surfaceFormat() optional for Studio3D
Our standard pattern is to expect applications to call

QSurfaceFormat::setDefaultFormat(Q3DS::surfaceFormat());

after creating the Q(Gui)Application, in order to ensure all OpenGL stuff
and Qt Quick uses the format the runtime has determined to be ideal.

With Studio3D however this is not always feasible - the Qt Quick scene
may be hosted in an application that knows nothing about 3D and surface
format gotchas. Therefore, make Studio3D force using the QQuickWindow's
format() in case it is different from the ideal one. (if there was a
setDefaultFormat(), as recommended, then the version/profile would match).

At the same time drop Q3DSEngine::surfaceFormat() as an API on Q3DSEngine.
There is no need for this (there's always the public
Q3DS::surfaceFormat()) and not having this makes future steps towards
getting rid of the globals easier.

Task-number: QT3DS-1645
Change-Id: I43e8c0da4aa17e03eb2c49755eb064dd2dc3d0ce
Reviewed-by: Andy Nichols <andy.nichols@qt.io>