Properly fix getStartAndEndTime
This is hopefully the the final iteration...
Always go through the property updates and get the values as needed,
expecting the caller to know/remember to apply the property changes
is just hazardous, so let's just do it in-place.
Instead of iterating over the same data several times (layers and nodes)
just do it in one go.

Change-Id: Ide3d84570d4765567ed300a1107b010bb3e10b82
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>