Doc: Fix navigation issues for online docs
In the sidebar, link to the main Qt 3D Studio docs in parent directory,
add QML/C++ reference links, and fix the rest of the links.

Add a relative URL for this project to make links to this project
point to a 'runtime' subdirectory.

Rename the landing page to qt3d-runtime-index.html - this forces qdoc
to display the sidebar also on the landing page.

Fix .qhp generation by using lowercase filename for .qdocconf, and
fix inconsistencies in .qdocconf variable names.

Change-Id: I02b054d5b39f95d50cd1e4d267edabff7138f010
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>