Add support for using Q3DSViewerSettings in Q3DSEngine
Q3DSEngine has a reference to the Q3DSViewerSettings for usage in
generating the viewport for the main presentation. This is used to make
sure that the scene is rendered with the correct scale mode and matte if
requested on the main presentation.

Adds Menu items for scale modes and matte to Q3DSViewer

Change-Id: I9f468c42fe846be8170dac01fcdf6d692aae19ab
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>