When no part in model mesh with sub-meshes use last instead of first
This resolves an issue with .mesh files generated by older versions of
Q3DS that stored revisions in the .mesh file. The models did not
explicitly specify a part number for the latest revision, so the oldest
revision would be used instead because that was part #1.

Now when a mesh source has no part number, we will by default use the
last mesh in the mesh list. This will still typically be #1 because non-
submesh meshes still will only have 1 part entry.

Task-number: QT3DS-1791
Change-Id: I8c9633c545c2f4f1fa325d9b0a28ab5981acd3b6
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>