Fix image probe FOV calculation
Fix FOV equations. Also, fix case for FOV of 180 degrees where image
lighting map was used in wrong orientation. With this fix, the negative z
axis passes through the center of the lighting map, with edges of the image
wrapped around the scene, i.e. with default camera settings the center of
image is in the direction of camera. This holds true for both cases of FOV = 180 and
FOV < 180 degrees.

This fix must be replicated on editor side for results to match.

Task-ID: QT3DS-154

Change-Id: Idb8071dcd74cc3d5b17c964239bec9a52b4fe6b8
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>