Provide shadow maps for custom materials as one param
because a QParameter with name[index] and a single value does
not work as expected for uniform arrays of samplers (or scalars - structs
are different however).

Qt3D expects name[0] and a list of values, and has no support for
mapping a single value with a specific index in the name to the
appropriate uniform location. This is due to how GL uniforms work and
it is silly but we can accommodate by building a QVariantList and using
a single QParameter per uniform.

Task-number: QT3DS-1840
Change-Id: I84d587087f4c7545e1c0d915a9e916a39f1336b7
Reviewed-by: Andy Nichols <andy.nichols@qt.io>