Move towards enabling dynamically adding and changing models
Add an 'object' command to the console and extend the scene-change
handler in scenemanager.

Models can now change the mesh (sourcepath) dynamically. (in theory)

Also add C++ default values to Q3DSTextNode. newObject<> uses the
in-source default values, the metadata is only applied when parsing from uip.

Problems with not refreshing due to layer caching are now fixed.

Dynamically adding a light should now rebuild buffers, shadows, etc. as
expected.

Also brings back toggling the console with backtick to the widgets mode
of the viewer.

Add also a simple QEntity graph viewer in the profileui, showing
entities, object names and components. This is very useful to verify if
submeshes appear and disappear as expected for instance.

Task-number: QT3DS-1178
Change-Id: Ieae38cfcb3a7eeba81a9bbcc0c76c99296655d5c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>