Revise camera visibility handling
The directional light shadow issues were caused by not having an active
camera selected for the layer.

Implement a long-standing todo by relying on the global (inherited)
visibility (that takes both eyeball and the slide visibility into
account, up the parent chain) when choosing the active camera for a
layer.

This needs two things:

Moving the choose-a-camera (and updating camera-dependent things) stuff
to the point where the globalVisibility is actually changed. This is done
by a simple handleNodeGlobalChange() "callback" that is invoked from
updateGlobals(). This may be useful for other purposes in the future
(for example, to implement special handling of globalOpacity == 0).

The other is to ensure updateGlobals() is called whenever the eyeball or
the visibility tag changes. For nodes with special handling (layer,
camera) this was not always in place but should be corrected now.

In addition, updateSubTree(m_scene) becomes syncScene() since it is not really
suitable to be called with any other object anyway. The pending
visibility processing is moved out to a new function since sometimes we
will want to invoke it directly instead of waiting for the next frame action.

Also fixes a "camara" typo.

Task-number: QT3DS-1947
Change-Id: I7be957a9a3dbfba76daa5d0d73461de32e098be8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>