Support executing console commands on startup
With Q3DS_BOOTSCRIPT=0 we'll look for an autoexec.drgscr once
the main presentation's scene has been finalized (meaning all
subpresentations are ready was well) and execute the commands
in it. When the value is > 0 it specifies the number of seconds
to delay the execution. The latter is important since with 0 the
commands are run before the aspect engine is created, while with
the delay it is possible to test the case when entities get created
or destroyed at a later time.

Change-Id: Ib24b8bc7944a53874b58ffcb41ccc7c9931477dd
Reviewed-by: Andy Nichols <andy.nichols@qt.io>