Fix premul alpha handling for images
Assume non-premultiplied alpha which is the common case for regular
image files (we do not explicitly convert to ARGB32_Premultiplied or
such and therefore the image data is non-premultiplied in most cases
(except of course when the user explicitly creates an image with premul
alpha but we cannot do much about that)).

When using subpresentations as texture maps, premultiplied alpha is
assumed to be true since this is what both the layer compositor and
Qt Quick output. This is the only case for a premultiplied image source
(for now?).

This cleans up a number of todo notes in the source.

Task-number: QT3DS-1982
Change-Id: I73578da11a6e393bcbe467fce883478f54b1f5ef
Reviewed-by: Andy Nichols <andy.nichols@qt.io>